+++ lib/activerecord/null.rb
@@ -15,25 +15,48 @@ module ActiveRecord
     # Define a Null class for the given class.
     #
     # @example
-    #   class Business
+    #   class Business < ApplicationRecord
     #     Null do
     #       def name = "None"
     #     end
     #   end
     #
     #   Business.null # => #<Business::Null:0x0000000000000000>
+    #   Business.null.name # => "None"
     #
-    def Null(klass = self, &)
+    #   class User < ApplicationRecord
+    #     Null([:name, :team_name] => "Unknown")
+    #   end
+    #
+    #   User.null.name # => "Unknown"
+    #   User.null.team_name # => "Unknown"
+    #
+    # @param inherit [Class] The class from which the Null object inherits attributes
+    # @param assignments [Array] The attributes to assign to the null object
+    def Null(inherit = self, assignments = {}, &)
+      if inherit.is_a?(Hash)
+        assignments = inherit
+        inherit = self
+      end
       null_class = Class.new do
         include ::ActiveRecord::Null::Mimic
-        mimics klass
+        mimics inherit
 
         include Singleton
       end
-      null_class.class_eval(&)
-      const_set(:Null, null_class)
+      null_class.class_eval(&) if block_given?
 
-      define_singleton_method(:null) { null_class.instance }
+      nil_assignments = inherit.attribute_names
+      if assignments.any?
+        assignments.each do |attributes, value|
+          nil_assignments -= attributes
+          null_class.define_attribute_methods(attributes, value:)
+        end
+      end
+      null_class.define_attribute_methods(nil_assignments)
+      inherit.const_set(:Null, null_class)
+
+      inherit.define_singleton_method(:null) { null_class.instance }
     end
 
     def self.extended(base)