+++ lib/activerecord/null/mimic.rb
@@ -17,6 +17,18 @@ module ActiveRecord
         def self.table_name = @mimic_model_class.to_s.tableize
 
         def self.primary_key = @mimic_model_class.primary_key
+
+        def self.define_attribute_methods(attribute_names, value: nil)
+          attribute_names.each do |attribute_name|
+            unless method_defined?(attribute_name)
+              if value.is_a?(Proc)
+                define_method(attribute_name, &value)
+              else
+                define_method(attribute_name) { value }
+              end
+            end
+          end
+        end
       end
 
       def mimic_model_class = self.class.mimic_model_class